requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
python-dotenv==1.0.1
requests==2.31.0
pandas>2.0.0
